moving service instance validation and service instance chooser to the AbstractLoadBalancingClient
